+++ oci_api/graph/driver.py
@@ -72,7 +72,8 @@ class Driver(metaclass=Singleton):
     def load_filesystem(self, filesystem_json, layer=None):
         filesystem_id = filesystem_json['id']
         log.debug('Loading filesystem (%s)' % filesystem_id)
-        filesystem = Filesystem(filesystem_id, layer)
+        container_id = filesystem_json.get('container_id', None)
+        filesystem = Filesystem(filesystem_id, layer, container_id)
         layer_json = filesystem_json.get('layer')
         if layer_json is not None:
             self.load_layer(layer_json, filesystem)
@@ -81,8 +82,10 @@ class Driver(metaclass=Singleton):
     def load_layer(self, layer_json, filesystem):
         layer_descriptor = Descriptor.from_json(layer_json['descriptor'])
         log.debug('Loading layer (%s)' % layer_descriptor.get('Digest').encoded())
+        diff_id = layer_json['diff_id']
+        size = layer_json['size']
         images = layer_json.get('images', [])
-        layer = Layer(layer_descriptor, filesystem, images)
+        layer = Layer(layer_descriptor, diff_id, filesystem, size, images)
         self.layers[layer.id] = layer
         for filesystem_json in layer_json.get('filesystems', []):
             self.load_filesystem(filesystem_json, layer)
@@ -91,7 +94,6 @@ class Driver(metaclass=Singleton):
         driver_path = pathlib.Path(oci_config['global']['path'])
         driver_file_path = driver_path.joinpath('driver.json')
         log.debug('Start saving driver file (%s)' % driver_file_path)
-
         if self.filesystems is None:
             raise OCIError('Driver filesystems is not initialized, can not save')
         filesystems_json = [
@@ -99,7 +101,6 @@ class Driver(metaclass=Singleton):
                 for filesystem in self.filesystems.values()
                     if filesystem.layer is None
         ]
-
         if not driver_path.is_dir():
             driver_path.mkdir(filesystems=True)
         with driver_file_path.open('w') as driver_file:
@@ -112,13 +113,19 @@ class Driver(metaclass=Singleton):
 
     def filesystem_to_json(self, filesystem):
         filesystem_json = {'id': filesystem.id}
+        if filesystem.container_id is not None:
+            filesystem_json['container_id'] = filesystem.container_id
         layer = self.get_child_layer(filesystem)
         if layer is not None:
             filesystem_json['layer'] = self.layer_to_json(layer)
         return filesystem_json
 
     def layer_to_json(self, layer):
-        layer_json = {'descriptor': layer.descriptor.to_dict()}
+        layer_json = {
+            'descriptor': layer.descriptor.to_dict(),
+            'diff_id': layer.diff_id,
+            'size': layer.size
+        }
         if len(layer.images) > 0:
             layer_json['images'] = layer.images
         filesystems = self.get_child_filesystems(layer)
@@ -131,6 +138,12 @@ class Driver(metaclass=Singleton):
             raise FilesystemUnknownException('There is no filesystem with id (%s)' % filesystem_id)
         return self.filesystems[filesystem_id]
 
+    def get_filesystem_by_container_id(self, container_id):
+        for filesystem in self.filesystems.values():
+            if filesystem.container_id == container_id:
+                return filesystem
+        raise FilesystemUnknownException('There is no filesystem mounted for container (%s)' % container_id)
+
     def get_child_filesystems(self, layer):
         return [ 
             filesystem 
@@ -142,20 +155,37 @@ class Driver(metaclass=Singleton):
         log.debug('Start creating filesystem')
         if layer is not None and layer.id not in self.layers:
             raise LayerUnknownException('There is no layer with id (%s)' % layer.id)
-        filesystem = Filesystem(generate_random_sha256(), layer)
-        filesystem.create()
+        filesystem = Filesystem.create(layer)
         self.filesystems[filesystem.id] = filesystem
         self.save()
         log.debug('Finish creating filesystem')
         return filesystem
 
+    def mount_filesystem(self, filesystem, container_id, path):
+        log.debug('Start mounting filesystem (%s) for container (%s)' % (filesystem.id, container_id))
+        if filesystem.id not in self.filesystems:
+            raise FilesystemUnknownException('Filesystem (%s) is not in driver, can not mount'
+                % filesystem.id)
+        filesystem.mount(container_id, path)
+        self.save()
+        log.debug('Finish mounting filesystem (%s) for container (%s)' % (filesystem.id, container_id))
+
+    def unmount_filesystem(self, container_id, remove=False):
+        filesystem = self.get_filesystem_by_container_id(container_id)
+        log.debug('Start unmounting filesystem (%s) for container (%s)' % (filesystem.id, container_id))
+        filesystem.unmount(container_id)
+        if remove:
+            self.remove_filesystem(filesystem)
+        self.save()
+        log.debug('Finish unmounting filesystem (%s) for container (%s)' % (filesystem.id, container_id))
+
     def remove_filesystem(self, filesystem):
         filesystem_id = filesystem.id
         log.debug('Start removing filesystem (%s)' % filesystem_id)
         if filesystem_id not in self.filesystems:
-            raise FilesystemUnknownException('Filesystem (%s) is not in distribution, can not remove'
+            raise FilesystemUnknownException('Filesystem (%s) is not in driver, can not remove'
                 % filesystem_id)
-        if self.get_child_layer(filesystem) is not None:
+        if filesystem.is_mounted() or self.get_child_layer(filesystem) is not None:
             raise FilesystemInUseException('Filesystem (%s) is in use, can not remove' 
                 % filesystem_id)
         filesystem.destroy()
@@ -170,6 +200,12 @@ class Driver(metaclass=Singleton):
             raise LayerUnknownException('There is no layer with id (%s)' % layer_id)
         return self.layers[layer_id]
 
+    def get_layer_by_diff_id(self, diff_id):
+        for layer in self.layers.values():
+            if layer.diff_id == diff_id:
+                return layer
+        raise LayerUnknownException('There is no layer with diff id (%s)' % diff_id)
+
     def get_child_layer(self, filesystem):
         for layer in self.layers.values():
             if layer.filesystem == filesystem:
@@ -181,15 +217,7 @@ class Driver(metaclass=Singleton):
         log.debug('Start creating layer from filesystem (%s)' % original_filesystem_id)
         if filesystem.id not in self.filesystems:
             raise FilesystemUnknownException('Unknown filesystem (%s)' % filesystem.id)
-        try:
-            layer = Layer.create(filesystem)
-            del self.filesystems[original_filesystem_id]
-        except FilesystemInUseException as e:
-            filesystem_id = e.args[0]
-            log.debug('Filesystem (%s) was already present' % filesystem_id)
-            self.remove_filesystem(filesystem)
-            filesystem = self.get_filesystem(filesystem_id)
-            layer = self.get_child_layer(filesystem)
+        layer = Layer.create(filesystem)
         self.layers[layer.id] = layer
         self.filesystems[layer.filesystem.id] = layer.filesystem
         self.save()
@@ -200,7 +228,7 @@ class Driver(metaclass=Singleton):
         layer_id = layer.id
         log.debug('Start removing layer (%s)' % layer_id)
         if layer_id not in self.layers:
-            raise LayerUnknownException('Layer (%s) is not in distribution, can not remove'
+            raise LayerUnknownException('Layer (%s) is not in driver, can not remove'
                 % layer_id)
         if len(layer.images) > 0:
             images = ', '.join(layer.images)
@@ -218,7 +246,7 @@ class Driver(metaclass=Singleton):
         layer_id = layer.id
         log.debug('Start adding image reference (%s) to layer (%s)' % (image_id, layer_id))
         if layer_id not in self.layers:
-            raise LayerUnknownException('Layer (%s) is not in distribution, can not add image ref'
+            raise LayerUnknownException('Layer (%s) is not in driver, can not add image ref'
                 % layer_id)
         layer.add_image_reference(image_id)
         self.save()
@@ -228,7 +256,7 @@ class Driver(metaclass=Singleton):
         layer_id = layer.id
         log.debug('Start removing layer (%s) reference to image (%s)' % (layer_id, image_id))
         if layer_id not in self.layers:
-            raise LayerUnknownException('Layer (%s) is not in distribution, can not add image ref'
+            raise LayerUnknownException('Layer (%s) is not in driver, can not add image ref'
                 % layer_id)
         layer.remove_image_reference(image_id)
         self.save()