+++ oci_api/graph/filesystem.py
@@ -20,27 +20,31 @@ from oci_api.util import id_to_digest
 
 log = logging.getLogger(__name__)
 
+def get_filesystem_class():
+    driver_type = oci_config['driver']['type']
+    if driver_type == 'zfs':
+        from .zfs_filesystem import ZFSFilesystem
+        return ZFSFilesystem
+    raise OCIError('Unsupported driver type (%s)' % driver_type)
+
+
 class Filesystem:
     @classmethod
     def __new__(cls, *args, **kwargs):
         if cls is not Filesystem:
             return super(Filesystem, cls).__new__(cls, *args, **kwargs)
-        driver_type = oci_config['driver']['type']
-        if driver_type == 'zfs':
-            from .zfs_filesystem import ZFSFilesystem
-            return super(Filesystem, cls).__new__(ZFSFilesystem)
-        raise OCIError('Unsupported driver type (%s)' % driver_type)
+        return super(Filesystem, cls).__new__(get_filesystem_class())
 
-    def __init__(self, id, layer=None):
+    @classmethod
+    def create(cls, layer):
+        filesystem_class = get_filesystem_class()
+        return filesystem_class.create(layer)
+
+    def __init__(self, id, layer, container_id):
         log.debug('Creating instance of %s(%s)' % (type(self).__name__, id))
         self.id = id
         self.layer = layer
+        self.container_id = container_id
 
-    @property
-    def digest(self):
-        return id_to_digest(self.id)
-
-    @property
-    def path(self):
-        return pathlib.Path(oci_config['global']['path'], 'filesystems', self.id)
-
+    def is_mounted(self):
+        return self.container_id is not None